refactored existing code.py into programming.py and one submodule per existing tag
